+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -15,6 +15,7 @@ import { default as reasoningServer } from "@app/lib/actions/mcp_internal_action
 import { default as dustAppServer } from "@app/lib/actions/mcp_internal_actions/servers/run_dust_app";
 import { default as searchServer } from "@app/lib/actions/mcp_internal_actions/servers/search";
 import { default as tablesQueryServer } from "@app/lib/actions/mcp_internal_actions/servers/tables_query/server";
+import { default as tablesQueryServerV2 } from "@app/lib/actions/mcp_internal_actions/servers/tables_query/server_v2";
 import { default as thinkServer } from "@app/lib/actions/mcp_internal_actions/servers/think";
 import { default as webtoolsServer } from "@app/lib/actions/mcp_internal_actions/servers/webtools";
 import type {
@@ -49,6 +50,8 @@ export async function getInternalMCPServer(
       return childAgentDebuggerServer();
     case "query_tables":
       return tablesQueryServer(auth, agentLoopRunContext);
+    case "query_tables_v2":
+      return tablesQueryServerV2(auth, agentLoopRunContext);
     case "primitive_types_debugger":
       return primitiveTypesDebuggerServer();
     case "think":
